Functions of craps and roulette games + created the customers and tables + added the inehiritance property and now the classes should be well declared +created a list of lists of player by table +simulated one round +returned and seted the best for players mainly for returning custommers +updated the wealth of the customer after each round +created the classes of employees : barmen and croupiers +assigned a croupier for each table inside the class of the tables +corrected a small error inside the class barmen and added a function which adds tips and added an ID attribute to the class croupier and one more small mistake

+++ tables.py
@@ -82,7 +82,7 @@ class Croupier(Employee):
         self.croupierID = croupierID
     def commission(self, partofwin):
         if partofwin > 0:
-            self.partofwin += int(partofwin) * 0.05
+            self.partofwin += float(partofwin) * 0.05
 
 class Barman(Employee):
     def __init__(self, wage=0, alcsales =0):
